+++ full-moon/src/ast/mod.rs
@@ -279,6 +279,74 @@ impl Default for TableConstructor {
     }
 }
 
+/// An anonymous function, such as `function() end`
+#[derive(Clone, Debug, Display, PartialEq, Node, Visit)]
+#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
+#[cfg_attr(not(feature = "luau"), display("{function_token}{body}"))]
+#[cfg_attr(
+    feature = "luau",
+    display("{}{}{}", join_vec(attributes), function_token, body)
+)]
+pub struct AnonymousFunction {
+    #[cfg(feature = "luau")]
+    attributes: Vec<LuauAttribute>,
+    function_token: TokenReference,
+    body: FunctionBody,
+}
+
+impl AnonymousFunction {
+    /// Returns a new AnonymousFunction
+    pub fn new() -> Self {
+        AnonymousFunction {
+            #[cfg(feature = "luau")]
+            attributes: Vec::new(),
+            function_token: TokenReference::basic_symbol("function"),
+            body: FunctionBody::new(),
+        }
+    }
+
+    /// The attributes in the function, e.g. `@native`
+    #[cfg(feature = "luau")]
+    pub fn attributes(&self) -> impl Iterator<Item = &LuauAttribute> {
+        self.attributes.iter()
+    }
+
+    /// The `function` token
+    pub fn function_token(&self) -> &TokenReference {
+        &self.function_token
+    }
+
+    /// The function body, everything except `function` in `function(a, b, c) call() end`
+    pub fn body(&self) -> &FunctionBody {
+        &self.body
+    }
+
+    /// Returns a new AnonymousFunction with the given attributes (e.g. `@native`)
+    #[cfg(feature = "luau")]
+    pub fn with_attributes(self, attributes: Vec<LuauAttribute>) -> Self {
+        Self { attributes, ..self }
+    }
+
+    /// Returns a new AnonymousFunction with the given `function` token
+    pub fn with_function_token(self, function_token: TokenReference) -> Self {
+        Self {
+            function_token,
+            ..self
+        }
+    }
+
+    /// Returns a new AnonymousFunction with the given function body
+    pub fn with_body(self, body: FunctionBody) -> Self {
+        Self { body, ..self }
+    }
+}
+
+impl Default for AnonymousFunction {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
 /// An expression, mostly useful for getting values
 #[derive(Clone, Debug, Display, PartialEq, Node)]
 #[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
@@ -315,8 +383,8 @@ pub enum Expression {
     },
 
     /// An anonymous function, such as `function() end`
-    #[display("{}{}", _0.0, _0.1)]
-    Function(Box<(TokenReference, FunctionBody)>),
+    #[display("{_0}")]
+    Function(Box<AnonymousFunction>),
 
     /// A call of a function, such as `call()`
     #[display("{_0}")]
@@ -1620,8 +1688,20 @@ impl Assignment {
     not(feature = "luau"),
     display("{local_token}{function_token}{name}{body}")
 )]
-#[cfg_attr(feature = "luau", display("{local_token}{function_token}{name}{body}"))]
+#[cfg_attr(
+    feature = "luau",
+    display(
+        "{}{}{}{}{}",
+        join_vec(attributes),
+        local_token,
+        function_token,
+        name,
+        body
+    )
+)]
 pub struct LocalFunction {
+    #[cfg(feature = "luau")]
+    attributes: Vec<LuauAttribute>,
     local_token: TokenReference,
     function_token: TokenReference,
     name: TokenReference,
@@ -1632,11 +1712,19 @@ impl LocalFunction {
     /// Returns a new LocalFunction from the given name
     pub fn new(name: TokenReference) -> Self {
         LocalFunction {
+            #[cfg(feature = "luau")]
+            attributes: Vec::new(),
             local_token: TokenReference::basic_symbol("local "),
             function_token: TokenReference::basic_symbol("function "),
             name,
             body: FunctionBody::new(),
         }
+    }
+
+    /// The attributes in the function, e.g. `@native`
+    #[cfg(feature = "luau")]
+    pub fn attributes(&self) -> impl Iterator<Item = &LuauAttribute> {
+        self.attributes.iter()
     }
 
     /// The `local` token
@@ -1657,6 +1745,12 @@ impl LocalFunction {
     /// The name of the function, the `x` part of `local function x() end`
     pub fn name(&self) -> &TokenReference {
         &self.name
+    }
+
+    /// Returns a new LocalFunction with the given attributes (e.g. `@native`)
+    #[cfg(feature = "luau")]
+    pub fn with_attributes(self, attributes: Vec<LuauAttribute>) -> Self {
+        Self { attributes, ..self }
     }
 
     /// Returns a new LocalFunction with the given `local` token
@@ -1992,8 +2086,13 @@ impl FunctionName {
 #[derive(Clone, Debug, Display, PartialEq, Node, Visit)]
 #[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
 #[cfg_attr(not(feature = "luau"), display("{function_token}{name}{body}"))]
-#[cfg_attr(feature = "luau", display("{function_token}{name}{body}"))]
+#[cfg_attr(
+    feature = "luau",
+    display("{}{}{}{}", join_vec(attributes), function_token, name, body)
+)]
 pub struct FunctionDeclaration {
+    #[cfg(feature = "luau")]
+    attributes: Vec<LuauAttribute>,
     function_token: TokenReference,
     name: FunctionName,
     body: FunctionBody,
@@ -2003,10 +2102,18 @@ impl FunctionDeclaration {
     /// Creates a new FunctionDeclaration from the given name
     pub fn new(name: FunctionName) -> Self {
         Self {
+            #[cfg(feature = "luau")]
+            attributes: Vec::new(),
             function_token: TokenReference::basic_symbol("function "),
             name,
             body: FunctionBody::new(),
         }
+    }
+
+    /// The attributes in the function, e.g. `@native`
+    #[cfg(feature = "luau")]
+    pub fn attributes(&self) -> impl Iterator<Item = &LuauAttribute> {
+        self.attributes.iter()
     }
 
     /// The `function` token
@@ -2022,6 +2129,12 @@ impl FunctionDeclaration {
     /// The name of the function
     pub fn name(&self) -> &FunctionName {
         &self.name
+    }
+
+    /// Returns a new FunctionDeclaration with the given attributes (e.g. `@native`)
+    #[cfg(feature = "luau")]
+    pub fn with_attributes(self, attributes: Vec<LuauAttribute>) -> Self {
+        Self { attributes, ..self }
     }
 
     /// Returns a new FunctionDeclaration with the given `function` token